requirements.txt
annotated-types==0.7.0
anyio==4.12.1
argcomplete==3.6.3
attrs==25.4.0
Authlib==1.6.6
azure-cosmos>=4.9.0,<5
beartype==0.22.2
cachetools==6.2.6
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.1.8
cloudpickle==3.1.2
colorama==0.4.6
cryptography==46.0.4
cyclopts==3.24.0
diskcache==5.6.3
dnspython==2.7.0
docstring_parser==0.17.0
docutils==0.22.4
email-validator==2.3.0
exceptiongroup==1.3.1
fakeredis==2.33.0
fastmcp==2.12.5
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
idna==3.11
importlib_metadata==8.7.1
jaraco.classes==3.4.0
jaraco.context==6.1.0
jaraco.functools==4.4.0
jsonref==1.1.0
jsonschema==4.26.0
jsonschema-path==0.3.4
jsonschema-specifications==2025.9.1
keyring==25.7.0
lupa==2.6
markdown-it-py==4.0.0
mcp==1.16.0
mdurl==0.1.2
more-itertools==10.8.0
openai==2.7.1
openapi-pydantic==0.5.1
opentelemetry-api==1.39.1
opentelemetry-exporter-prometheus==0.60b1
opentelemetry-instrumentation==0.60b1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
packaging==26.0
pathable==0.4.4
pathvalidate==3.3.1
pipx==1.8.0
platformdirs==4.5.1
prometheus_client==0.24.1
py-key-value-aio==0.3.0
py-key-value-shared==0.3.0
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
pydocket==0.16.6
Pygments==2.19.2
PyJWT==2.10.1
pyperclip==1.11.0
python-dotenv==1.2.1
python-json-logger==4.0.0
python-multipart==0.0.22
PyYAML==6.0.3
redis==7.1.0
referencing==0.36.2
requests==2.32.5
rich==14.3.1
rich-rst==1.3.2
rpds-py==0.30.0
shellingham==1.5.4
sortedcontainers==2.4.0
sse-starlette==3.2.0
starlette==0.52.1
typer==0.21.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
userpath==1.9.2
uv==0.9.27
uvicorn==0.40.0
watchfiles==1.1.1
websockets==16.0
workos==5.40.0
wrapt==1.17.3
zipp==3.23.0
